langgraph/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
langgraph==0.0.26
langchain==0.1.4
langchain-openai==0.0.5
python-dotenv==1.0.0
PyPDF2==3.0.1
python-multipart==0.0.6